Isolate classes from the test file
