+++ src/utilities/localStorage.js
@@ -1,0 +1,18 @@
+class LocalStorage {
+  constructor() {
+    this.list = [];
+  }
+  setDataLocalStorage(item) {
+    this.list = item;
+  }
+
+  getDataLocalStorage() {
+    return this.list;
+  }
+
+  clearDataLocalStorage() {
+    this.list = [];
+  }
+}
+
+module.exports = LocalStorage;
