+++ src/utilities/task.js
@@ -1,0 +1,9 @@
+class Task {
+  constructor(description, status, index) {
+    this.description = description;
+    this.status = status;
+    this.index = index;
+  }
+}
+
+module.exports = Task;
